+++ crates/ruma-events/src/room/encryption.rs
@@ -20,6 +20,12 @@ pub struct RoomEncryptionEventContent {
     /// Must be `m.megolm.v1.aes-sha2`.
     pub algorithm: EventEncryptionAlgorithm,
 
+    /// Whether state events should be encrypted alongside message-like events.
+    #[cfg(feature = "unstable-msc3414")]
+    #[serde(default, skip_serializing_if = "ruma_common::serde::is_default")]
+    #[serde(rename = "io.element.msc3414.encrypt_state_events")]
+    pub encrypt_state_events: bool,
+
     /// How long the session should be used before changing it.
     ///
     /// `uint!(604800000)` (a week) is the recommended default.
@@ -36,7 +42,13 @@ pub struct RoomEncryptionEventContent {
 impl RoomEncryptionEventContent {
     /// Creates a new `RoomEncryptionEventContent` with the given algorithm.
     pub fn new(algorithm: EventEncryptionAlgorithm) -> Self {
-        Self { algorithm, rotation_period_ms: None, rotation_period_msgs: None }
+        Self {
+            algorithm,
+            #[cfg(feature = "unstable-msc3414")]
+            encrypt_state_events: false,
+            rotation_period_ms: None,
+            rotation_period_msgs: None,
+        }
     }
 
     /// Creates a new `RoomEncryptionEventContent` with the mandatory algorithm and the recommended
@@ -48,8 +60,19 @@ impl RoomEncryptionEventContent {
         // Defaults defined at <https://spec.matrix.org/latest/client-server-api/#mroomencryption>
         Self {
             algorithm: EventEncryptionAlgorithm::MegolmV1AesSha2,
+            #[cfg(feature = "unstable-msc3414")]
+            encrypt_state_events: false,
             rotation_period_ms: Some(uint!(604_800_000)),
             rotation_period_msgs: Some(uint!(100)),
         }
+    }
+
+    /// Enable encrypted state as specified in [MSC3414][msc].
+    ///
+    /// [msc]: https://github.com/matrix-org/matrix-spec-proposals/pull/3414
+    #[cfg(feature = "unstable-msc3414")]
+    pub fn with_encrypted_state(mut self) -> Self {
+        self.encrypt_state_events = true;
+        self
     }
 }